requirements.txt
selenium==2.53.6
bunch==1.0.1
nose==1.3.7
nose-parameterized-0.6.0